Here, cost functions were optimized via the optimizer called hyperopt  to produce the highest possible PoA.
These files correspond to the description of various networks on the paradox of Braess, both single and bi-commodity. The difference to the descriptions in the parent folder is that, here, the cost functions were optimized via the optimizer called hyperopt (see https://github.com/hyperopt/hyperopt) in order to produce the highest possible price of anarchy (PoA).